Remove unified memory
Unified memory is a like a mythological siren.  It seduces you with
promises of convenience and improved programmability, and then
dashes you against the jagged rocks of performance penalties.

Actually, I don't know yet if the gnina performance regression was
due to unified memory, but there were a lot of obviously unneeded
transfers going on.