Collate function for MolDataset
The static method `MolDataset.collateMolDataset` can be used as the
collate_fn within `torch.utils.data.DataLoader`. This will return: lengths,
centers, coordinates, types, radii and labels of all of the examples in
a batch. Length corresponds to the number of atoms in example as each
example will be padded to the length of the example with the most atoms
in the batch. This allows the use of `gridmaker.forward` on the gpu and
using molgrid with DataLoader and `torch.utils.data.distributed.DistributedSampler`.